Posts info in a single API call
A single SQL request to retrieves all necessary information about posts.
The comments are retrieved with a jointure.
The username are integrated
into the data sent, rather than the user id.

This allows to send a complete response with comments and username
already integrated. The frontend will no longer need to make several
calls to the API and deal with the complexity of the database. This is
also more efficient.

Due to this breaking change, adjustment were made to the Vue app to deal
with the nested response sent by the API.
The changes made to the API also allows to delete some helpers functions from the frontend.

Also, a button allows to display the comments if needed.
However, all the comments have already been retrieved and are already there.

This should change in a upcoming patch.