Retrieving only last comment
GetAll posts route will now retrieve only the last comment if there is
one.
The total number of comments will also be included in the API response.
Another route is drafted to retrieve the rest of the comments by post
id.

In the frontend, if there are more than one comment, a button will allows to fetch the rest of them.
Necessary changes made to types and helpers.