Routing depending on authentication
If user is not logged in, they will be redirected to the logging page.
If they are not logged in but they are on the loggin or register page,
nothing will happen.

User can logout, which will remove the token. They will be brought back
to the login or register page unless they are already on them.