Dealing with user id & Refactoring
The newly created user id is passed from back to front end in the response.
It is stored in Vuex store and in the token. In the authentication middleware, if a user id is requested in a route, it is verified against the one stored in the token.
This will make sure that the user cannot send a request with someone else's user id.

Also :
* Heavy renaming
* The backend deals with the formatting of the responses instead of the
  front.