Adding admin status
Database now has a admin column on Users table.
It represents a boolean.

This information is returned when loggin in a user.
It is then store in Vuex frontend side.

A `delete` button is added on posts and comments whether user is admin or not.
For now it does not do anything. In another patch, a API route will be
created to handle deletion of posts and comments.

This patch also include some renaming and typing.