Add image feature
This patch implements the possibility for the user to add an image to
their post.
The user can do so by adding the URL of an image (jpg, gif, png).

The URL is being checked in the backend (extension and https).

The image URL is stored in a different table and deleted when the post
is deleted.