Adding README.md
The project's README with detailed steps to run the Groupomania app
(database, backend and frontend).

A groupomania_bdd.sql file is now available to get started with nice
database data.

Also deleting unused files for now. We should add unit test later
though.

For now, I ignore ESLint to release the project. But it should be
enabled and properly configured in a later patch.