Building for GitHub page with a GitHub action
This patch adds a .yml file which will be read as a GitHub Action.
The project will be automanically built when pushing to GitHub and the distribution file content will be stored in
the gh-page branch.
This branch and its content are used to set up a GitHub page, so that the app can be used at `ccr-g.github.io/7_09042021`.
It is also a way to keep the main branch clean from distribution files.

Also moving from router history mode to hash mode which is the
most straightforward solution to test the app for now.
That way, reloading the page won't trigger an error anymore, as well as
reaching directly for a particular page.